requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pillow>=10.0.0
fastapi
uvicorn
python-multipart
timm>=0.9.0
albumentations>=1.3.0
seaborn>=0.12.0
pyyaml>=6.0
tensorboard>=2.13.0
